v0.1.1 to Beta

+++ src/components/Tomato.jsx
@@ -1,62 +1,43 @@
-import React, {useState} from "react";
+import React from "react";
 import styled from "styled-components";
 import relativeTime from 'dayjs/plugin/relativeTime'
 import 'dayjs/locale/zh-cn'
 import dayjs from 'dayjs';
 import TextareaDebounced from '../components/TextareaDebounced';
+import {PaddingRow} from "../common/style";
 
 dayjs.extend(relativeTime);
 dayjs.locale('zh-cn');
 
-const TomatoRowGroup = styled.div`
+const TomatoRowGroup = styled(PaddingRow)`
  
 `;
 
 const TomatoRow = styled.div`
-  cursor: pointer;
-  padding: 0.3rem 7vw;
-  font-size: 0.7rem;
-  color: ${window.ttnoteThemeLight.textColorDesc};
-  &:active {
-    background-color: #ECECEC;
-  }
-  @media (min-width: 576px) {
-    &:active {
-      background-color: transparent;
-    }
-    padding: 0.3rem 8vw;
-  }
-  
   display: flex;
-  //grid-template-columns: repeat(20, 1fr);
-  //grid-template-rows: 1fr;
-  //justify-items: end;
   align-items: center;
-  justify-content: space-between;
-`;
-
-const Sequence = styled.div`
-  //margin-right: 1em;
-  flex: 0 0 1.8rem;
-  text-align: center;
-  //grid-area: 1 / 1 / 2 / 2;
-  //justify-self: center; 
+  color: ${window.ttnoteThemeLight.textColorDesc};
+  font-size: 0.7rem;
 `;
 
 const MinutesCell = styled.div`
-  //margin-right: 1em;
-  //flex: 2;
-  //grid-area: 1 / 5 / 2 / 8;
   flex: 0 0 2rem;
-  text-align: end;
+  @media (min-width: 576px) {
+    flex: 0 0 3rem;
+  }
 `;
 
 const TimeCell = styled.div`
   //margin-right: 1em;
   //flex: 2;
   //grid-area: 1 / 11 / 2 / 16;
-  flex: 0 0 4rem;
   text-align: end;
+  flex: 0 0 3rem;
+  margin-right: 0.5rem;
+  @media (min-width: 576px) {
+    flex: 0 0 4rem;
+    margin-right: 1rem;
+  }
 `;
 
 const DeleteCell = styled.div`
@@ -65,27 +46,27 @@ const DeleteCell = styled.div`
   cursor: pointer;
 `;
 
-const CollapseCell = styled.div`
- //grid-area: 1 / 16 / 2 / 20;
-  flex: 0 0 2rem;
-  text-align: end;
-  cursor: pointer;
-`;
-
 const DescCell = styled.div`
-  font-size: 0.8rem;
-  color: ${window.ttnoteThemeLight.textColorTitle}; 
-  padding: 0.2rem 12vw;
+  flex: auto;
+  padding: 0.3rem;
+  border-radius: ${window.ttnoteThemeLight.borderRadiusPrimary};
+  font-size: 0.9rem;
+  color: ${props => props.todoDone ? window.ttnoteThemeLight.textColorDesc : window.ttnoteThemeLight.textColorTitle};
+  background-color: ${window.ttnoteThemeLight.bgColorPrimary};
+  //display: ${props => props.visible ? 'block' : 'none'};
+  display: flex;
+  align-items: center;
+  margin-left: 1.7rem; //1.4 + 0.3
+  //margin-right: 0.5rem; 
   @media (min-width: 576px) {
-    padding: 0.2rem 10vw;
+    //margin-right: 1rem; 
   }
-  display: ${props => props.visible ? 'block' : 'none'};
 `;
 
 function Tomato(props) {
-  const {sequence, tomato} = props;
-  const [tomatoDescShow, setTomatoDescShow] = useState(tomato.desc);
-  const fromNow = dayjs(tomato.created_at).fromNow();
+  const {tomato, todoDone} = props;
+  // const [tomatoDescShow, setTomatoDescShow] = useState(tomato.desc);
+  const fromNow = dayjs(tomato.createdAt).fromNow();
 
   const saveInfo = (value) => {
     const url = window.ttnote.baseUrl + '/tomatoes/' + tomato.id;
@@ -93,34 +74,33 @@ function Tomato(props) {
       method: 'PATCH',
       body: JSON.stringify({desc: value}),
     })
-    .then(res => {
-      console.log(res);
-    })
+      .then(res => {
+        console.log(res);
+      })
   };
 
   const handleTomatoDelete = (e) => {
     e.stopPropagation();
-    props.deleteTomato(tomato.todoId, tomato.id)
+    if (window.confirm('确定要删除吗')) {
+      props.deleteTomato(tomato.todoId, tomato.id)
+    }
   };
 
   return (
     <TomatoRowGroup>
-      <TomatoRow onClick={() => setTomatoDescShow(!tomatoDescShow)}>
-        <Sequence>{`${sequence}.`}</Sequence>
-        <MinutesCell>{`${tomato.minutes}分钟`}</MinutesCell>
+      <TomatoRow>
+        <DescCell todoDone={todoDone}>
+          <TextareaDebounced
+            defaultValue={tomato.desc}
+            saveInfo={saveInfo}
+          />
+        </DescCell>
         <TimeCell>{fromNow}</TimeCell>
-        <CollapseCell
-        >{tomatoDescShow ? '折叠' : '展开'}</CollapseCell>
-        <DeleteCell
-          onClick={handleTomatoDelete}
-        >删除</DeleteCell>
+        <MinutesCell>{`${tomato.minutes}分钟`}</MinutesCell>
+          <DeleteCell
+            onClick={handleTomatoDelete}
+          >删除</DeleteCell>
       </TomatoRow>
-      <DescCell visible={tomatoDescShow}>
-       <TextareaDebounced
-         defaultValue={tomato.desc}
-         saveInfo={saveInfo}
-       />
-      </DescCell>
     </TomatoRowGroup>
   )
 }
